Add GameState for holding game state
We store the data in an agent, pass around it's pid and rely on this
module for each update to it. State also contains some temporary extras
(nickname, token, navigator) that shouldn't really be here

For now only the first level of struct/map nesting is a struct
(%State{}), the nested children are just plain maps with string keys.