Add Client module with main logic
We connect to the websocket and spin up the GameState process, then we
use recursion to keep track of socket and state_pid. For each loop we
pause at Socket.Web.recv! which will wait until a message is received.

And because of that we a separate process for regularly running code
like determining where to drive and here Navigator.start comes in.

For each type of message we can receive from server, we have a handle
function specialized for that datatype. And for each of these we make
sure to call loop(socket, state_pid) at the end to continue. For state
updates we always delegate to GameState module and pass in state_pid.